feat(preset-jekyll): support configured post types
